Caching (#24)
Caching an object means that it is converted into a rasterized image and doesn't have to be completely redrawn on each tick of EaselJS. This is especially useful for shapes, containers, or very large images.

Caching works transparently to the developer who uses a component. Any changes to props that can change the drawing of the object will cause the cache to refresh. Automatic changes to sprites will cause the cache to refresh.

EaselCache is a mixin so it can be incorporated as desired into new Easel* components.

With caching in place, it is now possible to implement filters.